chore: move `DirectoryItem` to `containers`
