+++ app/containers/DirectoryItem/styles.ts
